refactor: delete unused files
