--- src/resources/boards/board.model.ts
@@ -1,12 +1,0 @@
-import { v4 as uuid } from 'uuid';
-import Column from './column.model';
-
-class Board {
-  constructor(
-    public id: string = uuid(),
-    public title: string,
-    public columns: Column[]
-  ) {}
-}
-
-export default Board;
