--- src/resources/entities/index.ts
@@ -1,4 +1,0 @@
-export { User } from './user';
-export { Board } from './board';
-export { Colum } from './colum';
-export { Task } from './task';
